+++ R/get_predictions_coxph.R
@@ -1,4 +1,4 @@
-get_predictions_coxph <- function(model, fitfram, ci.lvl, typical, fun, vcov.fun, vcov.type, vcov.args, condition, ...) {
+get_predictions_coxph <- function(model, fitfram, ci.lvl, typical, fun, vcov.fun, vcov.type, vcov.args, condition, interval, ...) {
   # does user want standard errors?
   se <- !is.null(ci.lvl) && !is.na(ci.lvl)
 
@@ -18,7 +18,7 @@ get_predictions_coxph <- function(model, fitfram, ci.lvl, typical, fun, vcov.fun
     )
 
   # did user request standard errors? if yes, compute CI
-  if (!is.null(vcov.fun)) {
+  if (!is.null(vcov.fun) || (!is.null(interval) && interval == "prediction")) {
     # copy predictions
     fitfram$predicted <- exp(prdat$fit)
 
@@ -32,7 +32,8 @@ get_predictions_coxph <- function(model, fitfram, ci.lvl, typical, fun, vcov.fun
         vcov.fun = vcov.fun,
         vcov.type = vcov.type,
         vcov.args = vcov.args,
-        condition = condition
+        condition = condition,
+        interval = interval
       )
 
     if (!is.null(se.pred)) {
@@ -46,6 +47,7 @@ get_predictions_coxph <- function(model, fitfram, ci.lvl, typical, fun, vcov.fun
 
       # copy standard errors
       attr(fitfram, "std.error") <- se.fit
+      attr(fitfram, "prediction.interval") <- attr(se.pred, "prediction_interval")
 
     } else {
       # CI