+++ tests/testthat/test-linear-models.R
@@ -31,6 +31,18 @@ if (suppressWarnings(
     ggpredict(fit, c("c12hour", "c161sex"), vcov.fun = "vcovHC", vcov.type = "HC1")
   })
 
+  test_that("ggpredict, lm-prediction-interval", {
+    pr <- ggpredict(fit, c("c12hour", "c161sex"), interval = "predict")
+    expect_equal(pr$conf.low[1], 27.43113, tolerance = 1e-4)
+    pr <- ggpredict(fit, c("c12hour", "c161sex"), interval = "conf")
+    expect_equal(pr$conf.low[1], 71.02894, tolerance = 1e-4)
+    pr <- ggpredict(fit, c("c12hour", "c161sex"), interval = "predict", vcov.fun = "vcovHC", vcov.type = "HC1")
+    expect_equal(pr$conf.low[1], 27.44084, tolerance = 1e-4)
+
+    ggpredict(fit, c("c12hour", "c161sex"), interval = "predict", ci.lvl = NA)
+    ggpredict(fit, c("c12hour", "c161sex"), interval = "conf", ci.lvl = NA)
+  })
+
   test_that("ggpredict, lm-noci", {
     ggpredict(fit, c("c12hour", "c161sex"), ci.lvl = NA)
   })